+++ src/Model/Validation/PostsCategoryValidator.php
@@ -12,21 +12,11 @@
  */
 namespace MeCms\Model\Validation;
 
-use MeCms\Validation\AppValidator;
+use MeCms\Validation\CategoryValidator;
 
 /**
  * PostsCategory validator class
  */
-class PostsCategoryValidator extends AppValidator
+class PostsCategoryValidator extends CategoryValidator
 {
-    /**
-     * Construct
-     */
-    public function __construct()
-    {
-        parent::__construct();
-
-        $this->requirePresence('title', 'create');
-        $this->requirePresence('slug', 'create');
-    }
 }